requirements.txt
yfinance>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
ta>=0.11.0
vaderSentiment>=3.3.2
fredapi>=0.5.0
praw>=7.7.0
newsapi-python>=0.2.7
jinja2>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
